requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
python-akismet==0.4.3
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
